Restructure cypress tests and add test for timeout case
