+++ game_frontend/cypress/integration/avatarWorker/codeExecution.spec.js
@@ -1,8 +1,9 @@
 /// <reference types="cypress" />
 
 import { DEFAULT_CODE } from '../../../src/redux/features/constants'
+import { testAvatarCode, checkComputedTurnResult } from '../../support/avatarCodeTester'
 
-describe('Cypress for aimmo', () => {
+describe('Avatar worker', () => {
   beforeEach(() => {
     cy.login()
     cy.deleteAllGames()
@@ -192,33 +193,22 @@ def next_turn(world_map, avatar_state):
   })
 })
 
-function testAvatarCode (avatarCode, expectedAction, expectedLog) {
-  cy.loadGameWithAvatarCode(avatarCode)
-
-  cy.fixture('gameState.json').then(gameState => {
-    cy.window()
-      .its('store')
-      .invoke('dispatch', { type: 'features/Game/SOCKET_GAME_STATE_RECEIVED', payload: gameState })
-  })
-
-  checkComputedTurnResult(expectedAction, expectedLog)
-}
-
-function checkComputedTurnResult (expectedAction, expectedLog) {
-  const getComputedTurnResult = win => {
-    const state = win.store.getState()
-    return state.action.avatarAction
+it('Gives a timeout message when the worker takes too long to respond', () => {
+  const avatarCode = {
+    code: `
+def next_turn(world_map, avatar_state):
+    while True:
+        pass
+    return MoveAction(direction.NORTH)
+`
   }
 
-  cy.window()
-    .pipe(getComputedTurnResult)
-    .should(computedTurnResult => {
-      expect(computedTurnResult).to.not.be.undefined
+  const expectedAction = {
+    action_type: 'wait'
+  }
 
-      const avatarAction = computedTurnResult.action
-      const avatarLog = computedTurnResult.log
+  const expectedLog =
+    'Hmm... your avatar is taking a long time to respond, is there a 🐛 in your code?'
 
-      expect(avatarAction).to.deep.equal(expectedAction)
-      expect(avatarLog).to.deep.equal(expectedLog)
-    })
-}
+  testAvatarCode(avatarCode, expectedAction, expectedLog)
+})
