+++ game_frontend/cypress/integration/avatarWorker/webWorker.spec.ts
@@ -1,0 +1,31 @@
+/// <reference types="cypress" />
+
+import { testAvatarCode } from '../../support/avatarCodeTester'
+
+describe('Avatar worker in a web worker', () => {
+  beforeEach(() => {
+    cy.login()
+    cy.deleteAllGames()
+    cy.addTestGame()
+  })
+
+  it('Gives a timeout message when the worker takes too long to respond', () => {
+    const avatarCode = {
+      code: `
+  def next_turn(world_map, avatar_state):
+      while True:
+          pass
+      return MoveAction(direction.NORTH)
+  `
+    }
+
+    const expectedAction = {
+      action_type: 'wait'
+    }
+
+    const expectedLog =
+      'Hmm... your avatar is taking a long time to respond, is there a 🐛 in your code?'
+
+    testAvatarCode(avatarCode, expectedAction, expectedLog)
+  })
+})
